feat: add bingeplan command
the bingeplan command accepts one argument, episodes,
which is a list of episodes that should be counted.
there are also two options, fill and regex.
fill determines to the format for ranges in the episodes argument:
by default: 1-2 -> 01, 02
with --fill=1: 1-2 -> 1, 2
regex determines the regex used when filtering episodes:
default value: " <i> .*\.mkv" (<i> will be replaced by the episode
number)